refactor: Flatten App\Exception
