+++ src/Exception/SchemaExecuteException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Exception\SqlRunner;
+namespace App\Exception;
 
 use Symfony\Component\HttpKernel\Exception\HttpException;
