+++ src/Twig/Components/Challenge/Tabs/UserQueryResult.php
@@ -7,8 +7,8 @@ namespace App\Twig\Components\Challenge\Tabs;
 use App\Entity\ChallengeDto\FallableSqlRunnerResult;
 use App\Entity\Question;
 use App\Entity\User;
-use App\Exception\SqlRunner\QueryExecuteException;
-use App\Exception\SqlRunner\SchemaExecuteException;
+use App\Exception\QueryExecuteException;
+use App\Exception\SchemaExecuteException;
 use App\Repository\SolutionEventRepository;
 use App\Service\QuestionSqlRunnerService;
 use App\Service\SqlRunnerComparer;
